LayoutTests/imported/w3c/web-platform-tests/offscreen-canvas/text/2d.text.baseline.invalid.worker-expected.txt


FAIL 2d assert_equals: ctx.textBaseline === 'top' (got bogus[string], expected top[string]) expected "top" but got "bogus"
